requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
email-validator==2.1.1
scikit-learn==1.7.1
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2
openpyxl==3.1.5
python-docx==1.1.2
reportlab==4.2.2
SpeechRecognition==3.10.4
gTTS==2.5.3
translate==3.6.1
Werkzeug==3.0.4
Jinja2==3.1.4
itsdangerous==2.2.0
MarkupSafe==2.1.5
threadpoolctl==3.5.0
joblib==1.4.2
